completed frontend wrks

+++ src/main/java/com/example/WhizzRecipe/repository/RecipeRepository.java
@@ -1,5 +1,6 @@
 package com.example.WhizzRecipe.repository;
 
+import java.math.BigInteger;
 import java.util.*;
 import com.example.WhizzRecipe.dto.Recipe;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -7,7 +8,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface RecipeRepository extends JpaRepository<Recipe, Long> {
+public interface RecipeRepository extends JpaRepository<Recipe, BigInteger> {
     List<Recipe> findByTitleContainingIgnoreCase(String title); // For searching by title
 
     List<Recipe> findByGlutenFreeTrue(); // For gluten-free recipes
